Version that works in usability lab
After we finished working on this project in class, we set
up a copy on a machine in a usability lab at UNL and configured
the project on that machine. Unfortunately, we didn't think ahead
very well, so the version on that machine was never put under
version control. In this commit, we've set up Git on that machine
and copied all of the changes from the usability lab machine into
the repo.

Unfortunately, I'm not entirely sure what everything that changed
was or why. There be dragons in this commit.

One particular gotcha in this commit is that the compiled .class
files are NOT in sync with the current .java source code files. My
next task is to make building and editing the program much easier,
then make sure that the build is either ignored or kept up-to-date.

+++ CodeMusic/cobertura_bin/NoStructure.java
@@ -1,6 +1,6 @@
 public class NoStructure {
 	public static void main(String[] args) {
-		double x = 1;
+		int x = 1;
 		x = x + 0;
 		x = x + 1;
 		x = x + 2;
